Merge bootcamp B3-solution
This will replace the android sample project
as the baseline with the bootcamp B3 solution.
This means, that all files from com/android/sample
were deleted and we now have com/github/se as the package name
(this must be changed in the future).

Copy some model files over to the new file structure.
Merge the config files that were touched by us,
such that usually our commits prevail.