Use android emulator action that installs the emulator
Currently CI is broken, due to the avdmanger (and therefore Android
emulator)
not being found.
This commit tries to fix this by replacing the current android emulator
initialization action with one that handles the SDK, avdmanager and
emulator setup.
(As specified by the comment above the task).
Also, uncomment enabling KVM hardware acceleration.