Refactor auth
Get rid of old AuthRepository
and its single implementation.
Instead we have now one centralized Auth class.
This auth class is handed to the relevant screens.
It wraps the CredentialManger,
so we don't have to interact with it inside of the app anymore.
This also simplifies the SignInViewModel significantly.
Rename Auth screen tag to SignIn for consistency.