Add listen methods to repos
Add `listen` and `listenAll` methods to event and assoc repos.
These can be used instead of `get` for live data.
Using listeners improves the offline experience,
as it can use cached data immediately.

The local implementation for listening is just to keep a list of
listeners
that is called when a crud operation is executed.

Register listeners in init methods and remove LaunchEffects,
since they have a bad influence on home and calendar screen,
when triggered from both.